scale-in checkpoint
