+++ planner/factories.go
@@ -3,6 +3,8 @@ package planner
 import (
 	simulationfactory "github.com/gardener/scaling-advisor/planner/simulation/factory"
 	simulatorfactory "github.com/gardener/scaling-advisor/planner/simulator/factory"
+	scaleincandidateselector "github.com/gardener/scaling-advisor/planner/simulator/scalein/candidateselector"
+	nodeutilizationcalculator "github.com/gardener/scaling-advisor/planner/utilization"
 	"github.com/gardener/scaling-advisor/planner/weigher"
 
 	plannerapi "github.com/gardener/scaling-advisor/api/planner"
@@ -14,13 +16,16 @@ var (
 
 // NewFactories returns an instance of plannerapi.Factories populated with implementation of factory facades.
 func NewFactories() plannerapi.Factories {
+	// TODO: construct NodeUtilizationCalculator and pass to scaleincandidateselector.New()
+	defaultNodeUtilizationCalculator := nodeutilizationcalculator.New()
+	defaultScaleInCandidateSelector := scaleincandidateselector.New(defaultNodeUtilizationCalculator)
 	return plannerapi.Factories{
-		Planner:                  &defaultFactory{},
-		Simulator:                simulatorfactory.New(),
-		Simulation:               simulationfactory.New(),
-		ScaleInCandidateSelector: scaleInCandidateSelectorFactory.New(),
-		// TODO: add ScaleInCandidateSelector and NodeUtilizationCalculator. ScaleInCandidateSelector will have NodeUtilizationCalculator embedded in it.
-		ResourceWeigher: weigher.New(),
+		Planner:                   &defaultFactory{},
+		Simulator:                 simulatorfactory.New(),
+		Simulation:                simulationfactory.New(),
+		NodeUtilizationCalculator: defaultNodeUtilizationCalculator,
+		ScaleInCandidateSelector:  defaultScaleInCandidateSelector,
+		ResourceWeigher:           weigher.New(),
 	}
 }
